requirements.txt
google-generativeai>=0.3.0
pyaudio>=0.2.13
python-dotenv>=1.0.0
numpy>=1.24.0
